implemented the decorator design pattern

+++ src/main/java/com/xebia/controller/RetailStoreController.java
@@ -1,5 +1,6 @@
 package com.xebia.controller;
 
+import java.math.BigDecimal;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -11,13 +12,13 @@ import com.xebia.service.BillService;
 @RestController
 public class RetailStoreController {
 
-	
+
 	@Autowired BillService billService;
-	
-	
+
+
 	@RequestMapping(value="/getBill/{id}", method=RequestMethod.GET)
-	public Double getBillDetails(@PathVariable("id") int id){
-		Double billAmount=billService.getBill(id);
+	public BigDecimal getBillDetails(@PathVariable("id") int id){
+		BigDecimal billAmount=billService.getBill(id);
 		return billAmount;
 	}
 }
