+++ src/main/java/com/xebia/service/BillService.java
@@ -1,10 +1,11 @@
 package com.xebia.service;
 
+import java.math.BigDecimal;
 
 public interface BillService {
 
 	/*
 	 *  Method to get bill details.
 	*/
-	public Double getBill(int id);
+	public BigDecimal getBill(int id);
 }
